requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
defusedxml==0.6.0
distlib==0.3.1
Django==3.0.8
django-allauth==0.42.0
django-ckeditor==6.0.0
django-crispy-forms==1.9.1
django-currentuser==0.5.0
django-js-asset==1.2.2
django-rosetta==0.9.4
django-translations==1.1.2
djangorestframework==3.12.4
filelock==3.0.12
idna==2.10
importlib-metadata==3.10.0
nodejs==0.1.1
oauthlib==3.1.0
optional-django==0.1.0
Pillow==8.1.0
polib==1.1.1
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.3
virtualenv==20.4.3
zipp==3.4.1

translations~=0.1.3
Pygments~=2.8.1
passlib~=1.7.4
